fix method naming
