+++ package.json
@@ -44,7 +44,7 @@
   "devDependencies": {
     "@types/fs-extra": "^8.0.1",
     "@types/jest": "^24.0.22",
-    "@types/node": "^13.5.0",
+    "@types/node": "^13.5.1",
     "@types/uuid": "^3.4.6",
     "jest": "^24.9.0",
     "ts-jest": "^24.1.0",
